discordListener/requirements.txt
discord
python-dotenv
PyNaCl
pydub
google-cloud-storage